Chore: Sonarr Upstream Sync - Batch 31 (September 2025) (#1192)
* Subtitles indexer flag to indicate BTN releases with subtitles (Sonarr/Sonarr@d6f265c7b)

* New: Georgian language (Sonarr/Sonarr@309b55fe3)

* Follow redirects when fetching Custom Lists (Sonarr/Sonarr@7fdc4d663)

* New: RQBit download client (Sonarr/Sonarr@37cb978f1)

* New: Changing icon during import to blue (Sonarr/Sonarr@478866b2b)

* New: Excluded Tags on Release Profile (Sonarr/Sonarr@644015105)

* New: Subtitles indexer flag to indicate releases with subtitles (Sonarr/Sonarr@64c421c18)

* Bump FluentMigrator to official 6.2.0 (Sonarr/Sonarr@3796c9e30)

* Bump System.Data.SQLite to official 2.0.2 (Sonarr/Sonarr@7dac00d5a)

* New: Retry SQLite writes for database is locked errors (Sonarr/Sonarr@2e1289b92)

* Add missing word to 'AddNewSeriesHelpText' English translation (Sonarr/Sonarr@15e935060)

* Fix clearing pending changes for First Run (Sonarr/Sonarr@c4978022e)

* New: Switch theme automatically on system change (Sonarr/Sonarr@4904e8588)

* Attempt to remove pid file only if config folder exists (Sonarr/Sonarr@8bab0a06d)

* Switch HttpProxySettingsProviderFixture to test cases (Sonarr/Sonarr@b1f77007d)

* Fixed: Prevent modals from overflowing screen width (Sonarr/Sonarr@6c581b7e3)

* Fixed: Incorrectly parsing English subtitles as English audio (Sonarr/Sonarr@224e74605)

* New: Parse 'Por' as Portuguese (Sonarr/Sonarr@a4e3be721)

* Bump coverlet.collector to official 6.0.4 (Sonarr/Sonarr@051451eb2)

* Avoid rewriting file names in builds (Sonarr/Sonarr@0a61e66ef)

---------

+++ src/NzbDrone.Common/EnvironmentInfo/AppFolderFactory.cs
@@ -166,7 +166,7 @@ namespace NzbDrone.Common.EnvironmentInfo
 
         private void RemovePidFile()
         {
-            if (OsInfo.IsNotWindows)
+            if (OsInfo.IsNotWindows && _diskProvider.FolderExists(_appFolderInfo.AppDataFolder))
             {
                 _diskProvider.DeleteFile(Path.Combine(_appFolderInfo.AppDataFolder, "whisparr.pid"));
             }
